setup the automatic publishing and cleared the structure of shapes package.
