+++ gbox/shapes/misc_shapes.py
@@ -1,0 +1,12 @@
+
+
+class Capsule(TopologicalClosedShape2D):
+    pass
+
+
+class CShape(TopologicalClosedShape2D):
+    pass
+
+
+class NLobeShape(TopologicalClosedShape2D):
+    pass
